// src/modules/api/directories/interfaces/interests.interface.ts
export enum InterestEnumType {
  clubs = 'Бары и клубы',
  sport = 'Спорт, йога',
  art = 'Литература, живопись',
  cinema = 'Кинематограф',
  music = 'Музыка',
  events = 'Техно вечеринки, фестивали',
  politics = 'Политика, финансы',
  fashion = 'Мода',
  cooking = 'Кулинария',
}